feat: product on home page show based on [user default address, user current location, default store] in this order

+++ backend/src/service/store.service.ts
@@ -9,10 +9,14 @@ import { GetStoreByIdInput } from "../schema/store/GetStoreByIdSchema";
 import { UpdateStoreInput } from "../schema/store/UpdateStoreSchema";
 import { GetNearestStoreInput } from "../schema/store/GetNearestStoreSchema";
 import { getDistance } from "geolib";
+import { prisma } from "../lib/db/prisma";
+import { AppError } from "../error/AppError";
 
 export class StoreService {
   static async createStore(data: CreateStoreInput) {
     const { name, phone, coords, addressName, description, postCode } = data;
+
+    const isExist = await prisma.store.findFirst();
     return await StoreRepository.createStore({
       name,
       phone,
@@ -21,6 +25,7 @@ export class StoreService {
       addressName,
       description,
       postCode,
+      isDefault: isExist ? false : true,
     });
   }
 
@@ -34,26 +39,69 @@ export class StoreService {
     return await StoreRepository.getStoreByIdWithEmployee({ id });
   }
 
-  static async getStores() {
+  static async getStoresWithEmployeeCount() {
     return await StoreRepository.getStoresWithEmployeeCount();
   }
 
-  static async updateStore(data: UpdateStoreInput) {
-    const { id, name, lng, lat, description, addressName, phone, postCode } =
-      data;
-    const store = await StoreRepository.getStoreByIdWithCounts({ id });
-    if (!store) throw new NotFoundError("Store Not Found");
+  static async getStoresWithProducts() {
+    return await StoreRepository.getStoresWithProducts();
+  }
 
-    return await StoreRepository.updateStore({
+  static async updateStore(data: UpdateStoreInput) {
+    const {
       id,
       name,
-      latitude: lat,
-      longitude: lng,
-      addressName,
+      lng,
+      lat,
       description,
+      addressName,
       phone,
       postCode,
-    });
+      isDefault,
+    } = data;
+    const store = await StoreRepository.getStoreById({ id });
+    if (!store) throw new NotFoundError("Store Not Found");
+
+    if (isDefault) {
+      if (!isDefault) {
+        console.warn(
+          `User with id ${id} is trying to set default store to false`
+        );
+        throw new AppError({
+          message: "Internal Server Error",
+          statusCode: 500,
+        });
+      }
+      const defaultStore = await StoreRepository.getDefaultStore();
+      if (!defaultStore) {
+        console.error("Default store not found when updating store");
+        throw new AppError({
+          message: "Internal Server Error",
+          statusCode: 500,
+        });
+      }
+      await prisma.$transaction(async(tx) => {
+       await tx.store.update({
+          where: { id: defaultStore.id },
+          data: { isDefault: false },
+        });
+        return await tx.store.update({
+          where: { id },
+          data: { isDefault: true },
+        });
+      });
+    } else {
+      await StoreRepository.updateStore({
+        id,
+        name,
+        latitude: lat,
+        longitude: lng,
+        addressName,
+        description,
+        phone,
+        postCode,
+      });
+    }
   }
 
   static async deleteStoreById(data: DeleteStoreByIdInput) {
@@ -68,6 +116,15 @@ export class StoreService {
     }
     if (store.productStores > 0) {
       throw new ConflictError("Products still exist");
+    }
+
+    const defaultStore=await prisma.store.findFirst({
+      where:{
+        isDefault:true
+      }
+    })
+    if(defaultStore?.id===id){
+      throw new ConflictError("Default store cannot be deleted");
     }
 
     return await StoreRepository.deleteStoreById({ id });
